fix: 가로 스크롤 상에서 세로 스크롤이 안되던 문제 해결
- 네이밍 Horizontal -> Nested
- 의미없지만 fragment_home 필요없는 layout wrap 제거
